Base Scene ANIMAL UPDATE
Updated base scene with new animals. Created a tweaking scene for art team, for the purpose of testing art changes before hand, called "TweakScene_BGprefab_WG". In addition, created a static scene that will only be updated once a change is approved, and should be used as the prefab extracting scene for the game by DEV.  New animals are not final, therefore located in the "Assets/Art/Test Elements" folder. This folder will be used for inserting assets for the purpose of art tweaking and testing. Please use this folder only.